refactor: moving triggers to the corresponding package
