--- src/scripts/process/background/notify.js
@@ -1,7 +1,0 @@
-import {onInstalledNotifier, onUninstalledNotifier} from "../../units/events/notify.bg";
-
-export function notify() {
-  onInstalledNotifier();
-  onUninstalledNotifier();
-}
-
