+++ src/tab.js
@@ -1,7 +1,12 @@
-import {actionExtensionIconClicked, actionPageLoaded} from "./scripts/process/tab/triggers";
 // import infos from "./infos";
-
-tab();
+import {
+  detectPageLoad,
+  domainExportChecker,
+  domainLoadChecker
+} from "./scripts/units/processing/checker/domainChecker.all";
+import {launchExport} from "./scripts/units/processing/exportProcess";
+import {checkClickCountAndDisplayModal} from "./scripts/units/interface/modals/clickCount.all";
+import {uiEnhancer} from "./scripts/units/triggers/uiEnhancer";
 
 async function tab() {
   // console.log(infos.APP_MODE)
@@ -10,3 +15,21 @@ async function tab() {
     await chrome.storage.local.set({isInjecting: false});
   });
 }
+
+export async function actionExtensionIconClicked() {
+  const domainPage = await domainExportChecker();
+  if (domainPage === null) return;
+  launchExport(domainPage);
+  checkClickCountAndDisplayModal(domainPage);
+}
+
+export async function actionPageLoaded() {
+  const domain = await domainLoadChecker();
+  if (domain === null) return;
+  const htmlCheck = detectPageLoad(domain);
+  if (!htmlCheck) return;
+  // scrapOnLoadListener();
+  uiEnhancer(domain);
+}
+
+tab();
